Local cache shared across backends
This commit changes the local cache to be shared between all backends rather than being local to each backend.
As mentioned in Issue #3, this change is motivated by future plans to support multiple backends concurrently, and this change will ensure the ProxyStore cache size variable is still valid in the context of multiple backends (e.g., the max cache size should not scale by the number of backends).

Specific changes:
- The cache is now a top-level variable in proxystore.backend.store
- Removed CachedStore
- Refactored some names of the provided backend stores
- Added a reset() function to the cache (mostly for testing purposes)
- Updated test cases accordingly