Factories can optionally put objects in store
To clean up the process of manually proxying an object, instead of having to 1) place the object in the store with a key, 2) create a factory with the key, and 3) create a proxy with the factory, factories can now optionally take the object as an argument.
If the object is passed to the factory's constructor, the constructor places the object in the backend first.