ProxyStore Refactor. Closes #8
This commit includes a major rewrite of the ProxyStore design with the
primary goal of more clearly separating the Proxy and Factory building
blocks from the specific implementations for working with different
types of object stores.

Major changes:
- Changed the `backend` subpackage to `store` to clarify that use of
object stores is not needed to use the other ProxyStore functionality
(i.e., they are not "backends" for the whole package).
- Stores are now initialized via name using a single, unified
`init_stores` function and can be retrieved via a `get_store` function (Issue #8).
- Support for multiple initialized backends (Closes #3).
- `ps.to_proxy` has been removed in favor of a `Store.proxy()` method.
This more cleanly indicates what store a proxy will be associated with
and removes the need to have one monolithic function that has to grow to
support more proxy and store types (Issue #8).
- Factories for a specific Store implementation are now included in the
file for the Store implementation rather that in `proxystore.factory` to
clarify that they are not general-use factories.

Minor code changes:
- Explicitly define classes as abstract where relevant (e.g.,
the base `Store` and `Factory` classes are now abstract).
- Force keyword arguments where relevant.
- Move proxy utilities to `proxystore.proxy`.
- Remove `proxystore.utils.evict()` because that functionality is
somewhat unneeded with the `evict` flag to `Store.proxy()` and because
if needed in the future, it should go in the `proxystore.store`
subpackage (Closes #2).
- Changed custom `Factory` pickled to use `__getnewargs_ex__` for better
keyword argument support.
- Removed support for cache_size 0 from `LRUCache`.
- Updated all unit test and API docs according to these code changes.

Documentation changes:
- Rewrote most of Quick Start and Advanced for clarity with a focus on
delineating the Proxy and Factory building blocks from the object store
implementations.
- API docs now show `__init__` methods.
- API docs now show inherited methods.
- API docs generate in the source file order by default with some
manual overrides for specific classes to have the methods in
alphabetical order (mostly when the class has inherited members).

Misc. Changes:
- Updated examples to work with new API.
- Fixed bug in FuncX example run script not knowing the endpoint ID.
- Update `tag_and_release.sh` to check the version in the documentation is
correct as well.
- Add `make html` so building docs does not required changing
directories.