Proxy with custom factories. Closes #6, Closes #11
`Store.proxy()` now takes an optional `factory` parameter to specify
what type of factory to instantiate and pass to the proxy that is
eventually returned by the function.

`RedisStore` and `LocalStore` have the `factory` parameter defaults to
`RedisFactory` and `LocalFactory`, respectively.

Adding this functionality allows users to intercept and redefine factory
behavior as needed without making a new type of `Store`.
For example, the request functionality in Issue #6 can be solved by
creating a custom factory that extends `RedisFactory`.
The custom factory enforces `serialize=False` and intercepts the
`resolve()` function to deserialize the object returned by the store
with a custom deserialization function.