Added batch methods for set and proxy
This addition was entirely motivated by the desire to add multiple objects to a GlobusStore but only trigger a single Globus transfer.
For all other Store types, calling set() or proxy() in a loop is fine, but there can be huge speedups for the GlobusStore if a single tranfers is used for many objects.