Factory out proxy() into RemoteStore
Added kwargs property to proxystore.store.Store so that we can factor out the
proxy() implementations from each subclass of Store into RemoteStore.
In other words, just reducing the amount of redundant code.

This has the other side effect of easily exposing the kwargs needed to
reinitialize a store instance.