GlobusStore endpoints parameters supports dict
GlobusStore can now take endpoints as a dictionary and
GlobusStore.kwargs will return endpoints as a dictionary. This change
ensures that the kwargs are easily serializable in case they need to be
moved to another machine (as we found in the case with Colmena).
Added GlobusEndpoints.dict() method to convert and endpoints objects to
a dict.