Store subclasses now handled kwargs independently
Previously, a subclass of Store needed to know all the kwargs of its
superclasses which made adding new kwargs to superclasses cumbersome.
For example, in the previous commits, the stats kwarg was added but was
not added to the subclasses kwargs properties.
Now, subclasses create a dict of their kwargs then pass the dict to
their super to ensure each class handles it's own kwargs.