Change UUID to be specified by client
Previously, the signaling server provided a peering client with a
UUID, but this tightly couples the endpoint to the peering system and
prevents the endpoint from having a constant local state. For example,
an endpoint in SOLO mode needs a UUID but may not connect to the
signaling server.

In the future, a proper endpoint registration service that is separate
from the signaling server and performs auth is needed, and this change
will allow for more easily plugging in a registration service.