Fix integration test hanging and better timeouts
The endpoint integration tests were hanging sometimes on GitHub Actions.
Specifically, the endpoints were unable to connect to the signaling
server. I suspect this is because of limited cores on the GitHub Actions
workers so the signaling server process was not always finishing before
the endpoint processes. This commit adds a wait until connection to
signaling server is possible in the fixtures for those integration
tests. The integration tests also find an open port now as opposed to
using a static one.

Since we also use a "while true: connect else sleep" pattern in many of
the tests, a global pytest timeout is added of 30 seconds per test, a
timeout for each GitHub actions workflow of 10 minutes is added, and the
sleeps in the while true loops were reduced to 10ms to speed up the
tests.