Improve failed peer connection handling
Previously, when a peer connection failed (e.g., because the signaling
server did not know the request peer), the PeerConnection would just
raise the error returned by the signaling server causing the endpoint to
crash.

Now, if the signaling server sends an error during peer connection, the
PeerConnection sets the exception on a future. The PeerConnection.ready()
method will raise the exception when called. The PeerManager will now
check the healh of the connection by calling ready() prior to using the
connection for anything. If the PeerManager catches an error from the
PeerConnection, the PeerManager logs it, removes the PeerConnection, and
safely kills any tasks associated with the connection.

On the Endpoint side, errors that bubble up from the PeerConnection to
the PeerManager will be caught and reraised but as a PeerRequestError.
The Quart app will catch these errors and return 400 status codes.