Use JSON instead of pickle for P2P messages (#59)
Note this is the first of two commits necessary to finish #59.
Specifically this commit changes messages between peers (i.e.,
PeerConnections) and the signaling server to use JSON rather than
pickle. The second commit will change messages between endpoints
to also use JSON.

Overall, the goal is to use the minimum amount of pickling necessary
to avoid arbitrary code injection. Objects stored in ProxyStore will
of course always need to be pickled, but the control messages do not
so these control messages are the target of the changes.