Run mypy and fix typing in tests/
Previously, tests/ was excluded from mypy in the pre-commit config
because mypy was not following the configurations in setup.cfg. It turns
out mypy was not following the rules because there were some missing
__init__.py files in tests/.

This commit removes the exclusions of tests/, adds the missing
__init__.py files, updates the mypy config for tests/testing to allow
partially typed defs. Of course, these changes also exposed many typing
related issues in the test dir (and some in proxystore itself) so all of
these issues have been resolved.