Chunk large objects in endpoint requests (#57)
Quart's default max request body size is 16 MB. To get around this, the
default was changed to 1 GB even though that is not a lasting solution.
To properly resolve the issue, requests from EndpointStore to an
endpoint are chunked into 16 MB chunks and responses from an Endpoint
are also chunked into 16 MB chunks.