Update to Quart 0.18.0
The breaking change between 0.17.0 and 0.18.0 is related to logging. The
goal in ProxyStore is to use the ProxyStore default logger everywhere,
but the removal of the "quart.serving" logger in 0.18.0 makes this no
longer possible. Specifically, the change below makes it not currently
possible to change the "access logger" used by Hypercorn for logging.
The result is that /almost/ all Quart and Hypercorn logging is done to
the logger configured by ProxyStore except for the Hypercorn access log.

This should be address if there is every a change in the Quart/Hypercorn
logging so that we can ensure the access logs are propogated to the file
handler logger.

https://github.com/pallets/quart/commit/a54dab248714f08ff9f49acb32dd47834f86ddc1#diff-315b1bc58e6a870119bd7edb29330255e5eeecc0daa314d49a8d1bd05e5ce407L1346