Use PyTest tmp_path for tmp_dir fixture
Previously, the tmp_dir fixture just created a file in /tmp/, but it is
not always guarenteed that /tmp/ is available on every system so we are
switching to the Pytest tmp_path fixture for better compat.

I also tried using Python's tmpfile.TemporaryDirectory, but there were
issues with the temp directory getting cleaned up with tests that used
multiprocessing and coverage.