Add SSL/TLS support to signaling server
Adds optional SSL/TLS support to the signaling server via optional
certfile and keyfile parameters. Clients using `connect()` to connect to
a signaling server now specify the address with ws:// or wss:// and SSL
is enabled/disabled on the client side accordingly.

Note that the change in `connect()` requiring the address to start with
ws:// or wss:// makes this a breaking change.