Improve peer connection performance
- Doubled max chunk size (few percentage point improvement)
- Message sending now respects the bufferAmountLowThreshold
- Channel transport flushing no longer done on every send (as the aiortc
  example do not do this)
- Refactored channel events

Between this and the last commit, the testing/scripts/peer_bandwidth.py
benchmark improved from ~70 Mbps to ~300 Mbps. This is roughly the same
performance as aiortc's datachannel-filexfer example. Most of the
improvement is likely due to using uvloop in the test; however, I saw
greatly varying performance between Ubuntu and MacOS.