Add bytes to PeerManager.send + remove PeerMessage
Updated PeerManager send/recv to support bytes in addition to string
since PeerConnection has been upgraded to support both. Removed
PeerMessage has the information in the PeerMessage was mostly redundant
and the error handling contained in the PeerMessage should really be
handled at a level higher than the PeerManager since the PeerManager
just managers a set of PeerConnections which are fairly low level.

The Endpoint was updated accordingly with the changes to the
PeerManager.