Allow detaching an endpoint from the CLI (#128)
`proxystore-endpoint start` now defaults to spawning the endpoint as
a detached process with its PID recorded to a PID file in the endpoint
directory. The `--no-detach` flag can be used to run the endpoint
directly in the commands process.

`proxystore-endpoint stop` has been added.

`proxystore-endpoint list` now shows the status of each endpoint.