Restore cwd after temporary change in test
This test erroneously changed the cwd but did not change it back. The
side effect is that if coverage produces another coverage file after
this test runs, that coverage file will end up in the wrong place and
not be combined with the others that live in the original cwd.

This problem has not manifested because none of the tests that run after
this one use multiprocessing (which would result in more coverage files
being created).