Fix bug in WebsocketServer handler
The WebsocketServer handler would crash with a
"websockets.exceptions.ConnectionClosedOK: received 1000 (OK);
then sent 1000 (OK)" error in some cases in the test suite (I think it
was a race condition but I am unsure... may have also been related to
which event loop the server was running in). The effect was the server
was closing after one message which initially was probably not caught
because the WebsocketStore creates a new client connection for every
message. This fix is for the handler to loop over all incoming messages
on a connection.

See related:
https://websockets.readthedocs.io/en/stable/faq/server.html#why-does-the-server-close-the-connection-after-one-message