[BREAKING] Remove init_store/STORES enum (#147)
The STORES enum causes a lot of problems (#147) with imports because
every Store type must be imported when proxystore.store is imported.
Additionally, having the STORES enum gave the appearance that it was an
exhaustive list of available Store types (but it is not because some
require additional packages and user can build their own).

After removing STORES, having init_store did not make sense as almost
all of the logic in init_store was converting the enum type or string
(using the STORES enum) to the actually class to initialize the store
with. Thus, init_store was basically a wrapper around initializing the
store and calling register_store.

Since init_store returns Store[Any], it has the additional downside of
losing the specific type information on the Store. So the decision is
made to remove init_store entirely in favour of register_store.

I think in the long run this will make it more explicit to the user what
is happening and help with static analysis.