Fix UCX task destroyed warning
Previously, UCX tests when run with UCX (i.e., not mocked) would produce
many "Task was destroyed but it is pending!" warnings. This is because
the ucp package did not correctly clean up its asyncio stuff before the
event loop was closed by pytest.

Specifically, ucp did not close and await the asyncio_task handles in
ProgressTask (see:
https://github.com/rapidsai/ucx-py/blob/d88d9f015167a405b393fd0b8c128367e89cdfc6/ucp/continuous_ucx_progress.py#L9
). They use the __del__to cancel the tasks, but 1) __del__ is not
guarenteed to be called by Python and 2) the tasks still needs to be
awaited after cancel. The global context for ucp maintains a list of
these ProgressTask objects, hence the multiple warnings.
UCP also does not remove the readers they add to the event loop in the
BlockingMode variant of ProgressTask.

The solution is to wrap ucp.reset() with our own functionality that gets
the global context and inspects it to cancel/await all tasks and remove
the readers. (Note I had to add a reset_ucp() and reset_ucp_async()
variants because on the server side the resetting is done in an event
loop and on the client side it is done outside of the event loop.) Then
we call these reset methods after the program is done: 1) at the end of
the ucx_store fixture for client code and 2) at the end of
UCXServer.run() for server code.

For debugging and mocking purposes, I refactored out the launch server
and wait for server utilities into top level functions.