Remove coverage from dim integration tests
Coverage appears to be the cause of the segfaults when running the
integration tests for mochi/margo. For now, I've just removed coverage
from the py{version}-dim tox environments, and we can assume some
coverage safety from the fact the mocked versions of the tests have full
coverage.

I also made a few other changes:
- factored out the mock multiprocessing context manager
- removed some no longer needed pragma: no covers because the
  integration tests no longer use coverage