Fix long teardown in UCX test (#172)
For some reason, having an async test in this module was causing
teardown to take 10 seconds. Even just having a no-op async test

@pytest.mark.asyncio
async def test_nothing() -> None:
    pass

would cause the 10 second teardown on the final test executed. I could
not figure out the exact cause for this, so I modified the test to not
be an async test.