Improve MultiStore initialization
The initial implementation of MultiStore used a cumbersome
StorePolicyArgs TypedDict for initialization. This structure was used
because factories need to be able to reinitialize the Store from the
Store.kwargs. For the MultiStore, this means the factory need to
reinitialize the MultiStore and the sub-stores. The factory isn't
capable of doing this so that logic was put in the MultiStore
constructor.

This commit allows users to construct a MultiStore with a mapping of
either store names to policies or store instances to polices. This is
much more readable.

To still support factories reinitializing all the sub-stores, the
StorePolicyArgs functionality is still there and used by
MultiStore.kwargs. I.e., a MultiStore can still be initialized from a
list of StorePolicyArgs, but not StorePolicyArgs is hidden as
_StorePolicyArgs.