Collect stats on Connector of Store (#212)
With the move to the Connector model, the stat tracking on the Store
lost a lot of functionality. This commit restores previous functionality
and expands on it but tracking stats on the Store, Connector, and
Factory operations independently.

This does mean that the keys returned by `Store.stats()` have changed,
but they are more informative now.

Issue #212 was fixed in the process of adding new functionality, and the
docs guide/example was updated.