Support dormant connectors (#253)
A dormant connector is a connector maintained by a MultiConnector
instance, but the dormant connector itself is not initialized because it
is not "valid" on that processes. Validity is determined by the
`host_pattern` of the connector's corresponding policy.

This allows one to use a `MultiConnector` object on many processes
across multiple machines even though not all of the individual
connectors may be valid or initializable on any specific machine.

For example, consider a `MultiConnector` managing a `RedisConnector`
available a site A and an `EndpointConnector` available at sites A and
B. If the `MultiConnector` gets initialized on site B, it needs to skip
the initialization of the `RedisConnector` because it is not available
at this site. Thus, the `RedisConnector` is determined dormant.