Change Store.set() to Store.put()
This name change brings the `Store` in line with the rest of the package
with respect to the meanings of set and put. We use set to refer to an
operation which takes a key and a value and returns nothing, and put to
refer to an operation which takes a value and returns a unique key.