Provide spawn Margo server shortcut (#295)
By default, the MargoConnector will attempt an RPC to the address it
expects a server to be listening on. This "ping" is used by the
MargoConnector to determine if a new server needs to be spawned.

However, Margo is very slow to raise an error when you attempt to make
an RPC to an address without a server. This can take as much as 90
seconds and causes the test suite to be very slow.

This commit adds a force_spawn_server parameter to the MargoConnector
which is False by default (preserving the old behavior). If True, the
MargoConnector will not attempt the "ping" of a server and will instead
immediately spawn a server. This flag is useful when the caller knows
they intend to spawn a server such as in the test suite where we
intentionally spawn a couple servers in the integrations tests.