Handle non-proxiable types in Store (#311)
Some types (at the time, just the built-in constants [True, False,
None]) do not work well as proxies (see Issue #311). Since proxying
one of these constants is not the intended use of the Store, this commit
will raise a NonProxiableTypeError when proxy(), proxy_batch(), or
locked_proxy() are called on bool/None. The skip_nonproxiable flag as
been added to these methods which will silently return the same object
instead of proxying it if it would have otherwise raises a
NonProxiableTypeError.

The overall goal of the change is to prevent users from making proxies
of these built-in constants to avoid confusion (such as why
"Proxy(None) is None" is actaully False).

Note that a lot of attention was given to how MyPy infers the return
types of non-proxiable types. For the most part, MyPy will correctly
determine if the return type is a proxy or one of the non-proxiable
types. The only exception is proxy_batch() on a sequence of multiple
types will always say the return value is a list of proxies even though
in reality it may be a mix of proxies and non-proxies.