Reconnect client to relay server in background
The RelayServerClient will automatically reconnect on send/recv, but
this isn't helpful when the client is just in listening mode for new
peers. In that situation, the client will not reconnect to the relay
server because an action will never be performed which initiates that
reconnection. Now the RelayServerClient will spawn a background task
which waits for the current websocket to close and initiate the
reconnect.