Fix test where websocket not closed before check
The test_relay_server_auto_reconnection failed in CI because the
check `assert not client._websocket.closed` failed. We briefly
sleep before this check to ensure the internal mechanisms of the
websocket get a chance to handle the closure. This might have been
enough (though getting the test to reliably fail eludes me), so we
instead yield the event loop 10 times before checking.