Create RelayClient protocol for PeerManager (#360)
The `PeerManager` and `PeerConnection` classes now take a `RelayClient`
protocol rather than the `PeerManager`. This will allow using different
kinds of relay servers in the future (such as those planned in #278).

Because the `PeerManager` is no longer responsible for connecting to the
relay server, connection parameters (e.g., the name, uuid, etc) and
associated logic have been moved into the `RelayClient`. For example,
the `PeerManager` gets it's identity (name and uuid) from the `RelayClient`
it was intialized with.