Make Endpoint dependent on a PeerManager (#361)
Previously, an Endpoint would initialize it's own RelayClient and
PeerManager and required all of the parameters for the
RelayClient/PeerManager to be passed to the Endpoint constructor.
Now a dependency injection model is used such that an Endpoint is
(optionally) initialized with an already created PeerManager.

This is a followup to #363 which refactored the PeerManager and
relay server clients so that different RelayClient's could be
used with the PeerManager. Now that the PeerManager is initialized
outside of the Endpoint, one can also use different RelayClient's
with an Endpoint. This is the motivation for this change because in
the future we want to support Endpoints connecting to different
kinds of relay servers