Extend some connector types as deferrable (#393)
Note that not all connector implementations were extended to implement
the DeferrableConnector interface because some connector types cannot
support creating a key before the object exists. For example,
GlobusConnector needs to start the Globus Transfer before creating the
key, and MultiConnector needs to know the object to be able to route the
object to the correct connector.