Improve status codes of endpoint responses (#413)
Previously, all errors in the endpoint serving routes returned HTTP
status code 400. This is misleading so the error codes have been updated
to correctly reflect the type of error. Additionally, the endpoint
routes are now included in the API reference with their response types.

The proxystore.endpoint.client.get() function caught 400 errors and
assumed that meant the data requested was not available. This is not
always the case, so that function was updated to only catch 404 errors
and raise all other error code.