Add streaming prototype from summer project
This commit is to maintain the original interface worked on by
Naz1380 and valhayot in the streaming branch. Due to the limitations
outlined in Issue #442 and challenges with rebasing the original
streaming branch on to main, I've decided to start the implementation of
Issue #442 from scratch.

That said, I wanted to maintain the authorship and original interface in
the commit history so I've copied the main streaming.py file here. I've
left out the custom modifications to the Connector interfaces because
they are pretty bespoke to this specific interface.

The next commit will essentially be modifying the streaming interface
here to match that of Issue #442.

Co-authored-by: Naz1380 <he4687@wayne.edu>
Co-authored-by: valhayot <valeriehayot@gmail.com>