Remove custom topic closing in shims
A publisher shim to a third-party message broker should not really
be responsible to closing subscribers to the third-party message broker.
This is because it's reasonable to expect the publisher and subscribers
may come and go over time. This previous implementation detail of the
Publisher/Subscriber protocol made specific instances unnecessarily
coupled to each other.

In the future, we may move this funtionality up to the higher level
StreamProducer/StreamConsumer interfaces rather than required it to be
implemented in every shim.