Fix Store.put_batch called with no objects
This would happen when a StreamProducer is closed and there are no
objects in the current batch buffer. A lot of connector implementations,
such as the ones used to test the StreamProducer, are okay with
put_batch([]) but the RedisConnector raises an error here. Maybe in the
future we should support empty batches in put_batch() as a part of the
Connector protocol, but it is definitely not needed to call put_batch in
the StreamProducer in this case.