Move mypy plugin to proxystore.mypy_plugin (#544)
Having the proxystore.mypy plugin causes issues when loading the plugin
via a path (as opposed to module) as we do in pyproject.toml. Mypy loads
in the "mypy" module within "proxystore/" which overwrites the actual
mypy module which proxystore.mypy imports from.

While we could specify "proxystore.mypy" within the pyproject.toml and
this problem would go away (I also think it looks cleaner), this causes
problems with the mypy pre-commit hook. Pre-commit does not install the
local package so "proxystore.mypy" is not a valid import for mypy when
run within the pre-commit virtual environment.

Renaming proxystore.mypy to proxystore.mypy_plugin prevents the name
clash and will have to do.