Add cache_defaults flag to Proxy (#549)
The cache_defaults flag and target parameter, when combined, will
compute and cache the __class__ and __hash__ values of the proxy such
that the proxy can be hashed or used with isinstance() without needing
to resolve the proxy.

This is particularly important for certain serializers or workflow
systems (like Dask) that want to do some custom management of values
which could be proxies.