Prefix Proxy special attributes as __proxy_*
There's quite a few special attributes between the base Proxy and the
reference Proxy implementations. To keep it clear what is a special
Proxy attribute and an attribute of the target, I've changed all the
special Proxy attributes to be prefixed with __proxy_*.

This is similar to how pydantic v2 prefixes all the pydantic specific
attributes and methods with "model_". This is not a breaking change
because these are all internal implementation details of the Proxy.

There were a few places where __factory__ was accessed so I've added a
get_factory() function to make that the preferred way of getting the
factory.