Refactor stream interfaces
The StreamProducer can now work without a mapping of topics to Stores.
When a Store is not assigned to a topic, the stream object is sent
directly inside the event. Thus, there are now two new object events:
NewObjectEvent for when the object is included directly, and
NewObjectKeyEvent for when the object is located in a Store.

The StreamProducer/StreamConsumer now support both types of
Publisher/Subscriber pairs, event-based and message-based, depending on
what granularity the shim implementation wants access to.