Add topic to event types and remove _EventInfo
The _EventInfo became a little confusing with union types after adding
the different kinds of new object events, so I removed that in the
StreamConsumer. This did require moving some metadata from the
EventBatch to the specific event types, but this probably makes more
sense with some metadata only applying to certain types anyways.