Support static endpoint host and default to ip
A prior PR changed the default to fqdn, which resulted in a few users
having to revert back to IP host name resolution. This reverts the
default back to IP for broader compatibility.

I also added a "static" host type so users can force their endpoint
host to be at a certain address, rather than trying to infer the
correct host from the IP address or fqdn.