Improve clarity of proxy resolve errors
The stack trace of an exception raised in the factory of a proxy when
resolved is often confusing because it starts with "AttributeError:
'Proxy' object has no attribute '__proxy_target__'" and then the true
error is raised in the handling of the above error. This commit changes
the behavior to raise a ProxyResolveError which is raised from the error
caught from the factory call.