Change default buffering in file/globus connectors
The default buffering was set to 0 to disable buffering when opening
files so that file writes would appear immediately on the file system.
However, writes larger tha 2GB were getting silently truncated by the OS
as noted in Issue #704. To preserve the ability to disable buffering for
unit tests, a buffering parameter is added to the connectors.